+++ src/Quarry.Benchmarks/Benchmarks/ComplexJoinFilterPaginateBenchmarks.cs
@@ -31,7 +31,7 @@ public class ComplexJoinFilterPaginateBenchmarks : BenchmarkBase
             results.Add(new UserOrderDto
             {
                 UserName = reader.GetString(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
@@ -105,7 +105,7 @@ public class ComplexJoinFilterPaginateBenchmarks : BenchmarkBase
             results.Add(new UserOrderDto
             {
                 UserName = reader.GetString(0),
-                Total = reader.GetDecimal(1)
+                Total = reader.GetDouble(1)
             });
         }
         return results;
